data type changed to int from float

+++ control-flow/multiplication_table.py
@@ -1,4 +1,4 @@
-number = float(input("Enter a number to see its multiplication table:"))
+number = int(input("Enter a number to see its multiplication table:"))
 for x in range (10):
     result = number*x
     print(f"{number}*{x}={result}")
